Cleanup EdgeTxTheme class.
