+++ companion/src/firmwares/customisation_data.cpp
@@ -132,23 +132,6 @@ bool RadioLayout::CustomScreenData::isEmpty() const
   return strlen(layoutId) == 0;
 }
 
-void RadioTheme::init(const char* themeName, ThemeData& themeData)
-{
-  memset(&themeData, 0, sizeof(ThemeData));
-
-  memcpy(&themeData.themeName, themeName, THEME_NAME_LEN);
-
-  PersistentData& persistentData = themeData.themePersistentData;
-
-  persistentData.options[0].type =
-      zoneValueEnumFromType(ZoneOption::Type::Color);
-  setZoneOptionValue(persistentData.options[0].value, (unsigned int)WHITE);
-
-  persistentData.options[1].type =
-      zoneValueEnumFromType(ZoneOption::Type::Color);
-  setZoneOptionValue(persistentData.options[1].value, (unsigned int)RED);
-}
-
 void RadioLayout::init(const char* layoutId, CustomScreens& customScreens)
 {
   memset(&customScreens, 0, sizeof(CustomScreens));
