+++ radio/src/gui/colorlcd/layouts/layout_factory_impl.h
@@ -206,11 +206,6 @@ class BaseLayoutFactory: public LayoutFactory
 
     const uint8_t* getBitmap() const override { return bitmap; }
 
-    void drawThumb(BitmapBuffer * dc, uint16_t x, uint16_t y, uint32_t flags) const override
-    {
-      dc->drawBitmapPattern(x, y, bitmap, flags);
-    }
-
     const ZoneOption * getOptions() const override
     {
       return options;
