+++ radio/src/gui/colorlcd/page.cpp
@@ -23,6 +23,7 @@
 #include "mainwindow.h"
 #include "keyboard_base.h"
 #include "opentx.h"
+#include "theme.h"
 
 PageHeader::PageHeader(Page * parent, uint8_t icon):
   FormWindow(parent, { 0, 0, LCD_W, MENU_HEADER_HEIGHT }, OPAQUE),
