+++ radio/src/main.cpp
@@ -26,6 +26,7 @@
   #include "libopenui.h"
   #include "gui/colorlcd/LvglWrapper.h"
   #include "gui/colorlcd/view_main.h"
+  #include "theme.h"
 #endif
 
 #if defined(CLI)
